+++ app/controllers/oauth.ts
@@ -4,15 +4,15 @@
  * @namespace controllers/oauth
  */
 
-import * as sskts from '@motionpicture/sskts-domain';
-import * as bcrypt from 'bcryptjs';
+// import * as sskts from '@motionpicture/sskts-domain';
+// import * as bcrypt from 'bcryptjs';
 import * as createDebug from 'debug';
-import { Request } from 'express';
+// import { Request } from 'express';
 import * as jwt from 'jsonwebtoken';
-import * as request from 'request-promise-native';
+// import * as request from 'request-promise-native';
 
 // tslint:disable-next-line:no-require-imports no-var-requires
-const googleAuth = require('google-auth-library');
+// const googleAuth = require('google-auth-library');
 
 const debug = createDebug('sskts-api:controllers:oauth');
 // todo どこで定義するか
@@ -41,26 +41,26 @@ export interface ICredentials {
  * @param {Request} req リクエストオブジェクト
  * @returns {Promise<ICredentials>} 資格情報
  */
-export async function issueCredentials(req: Request): Promise<ICredentials> {
-    switch (req.body.grant_type) {
-        case 'urn:ietf:params:oauth:grant-type:jwt-bearer':
-            return await issueCredentialsByAssertion(req.body.assertion, req.body.scopes);
+// export async function issueCredentials(req: Request): Promise<ICredentials> {
+//     switch (req.body.grant_type) {
+//         case 'urn:ietf:params:oauth:grant-type:jwt-bearer':
+//             return await issueCredentialsByAssertion(req.body.assertion, req.body.scopes);
 
-        case 'client_credentials':
-            const scopes = (<string>req.body.scope).split(' ');
+//         case 'client_credentials':
+//             const scopes = (<string>req.body.scope).split(' ');
 
-            return await issueCredentialsByClient(req.body.client_id, req.body.client_secret, req.body.state, scopes);
+//             return await issueCredentialsByClient(req.body.client_id, req.body.client_secret, req.body.state, scopes);
 
-        // case 'password':
-        //     return await issueCredentialsByPassword(
-        //         req.body.client_id, req.body.state, req.body.username, req.body.password, req.body.scopes
-        //     );
+//         // case 'password':
+//         //     return await issueCredentialsByPassword(
+//         //         req.body.client_id, req.body.state, req.body.username, req.body.password, req.body.scopes
+//         //     );
 
-        default:
-            // 非対応認可タイプ
-            throw new Error(MESSAGE_UNIMPLEMENTED_GRANT_TYPE);
-    }
-}
+//         default:
+//             // 非対応認可タイプ
+//             throw new Error(MESSAGE_UNIMPLEMENTED_GRANT_TYPE);
+//     }
+// }
 
 /**
  * 主張から資格情報を発行する
@@ -69,20 +69,20 @@ export async function issueCredentials(req: Request): Promise<ICredentials> {
  * @param {string[]} scopes スコープリスト
  * @returns {Promise<ICredentials>} 資格情報
  */
-export async function issueCredentialsByAssertion(assertion: string, scopes: string[]): Promise<ICredentials> {
-    if (assertion !== process.env.SSKTS_API_REFRESH_TOKEN) {
-        throw new Error(MESSAGE_INVALID_ASSERTION);
-    }
+// export async function issueCredentialsByAssertion(assertion: string, scopes: string[]): Promise<ICredentials> {
+//     if (assertion !== process.env.SSKTS_API_REFRESH_TOKEN) {
+//         throw new Error(MESSAGE_INVALID_ASSERTION);
+//     }
 
-    // todo clientとstateどうするか
-    const payload = sskts.factory.clientUser.create({
-        client: '',
-        state: '',
-        scopes: scopes
-    });
+//     // todo clientとstateどうするか
+//     const payload = sskts.factory.clientUser.create({
+//         client: '',
+//         state: '',
+//         scopes: scopes
+//     });
 
-    return await payload2credentials(payload);
-}
+//     return await payload2credentials(payload);
+// }
 
 /**
  * クライアントIDから資格情報を発行する
@@ -91,215 +91,215 @@ export async function issueCredentialsByAssertion(assertion: string, scopes: str
  * @param {string[]} scopes スコープリスト
  * @returns {Promise<ICredentials>} 資格情報
  */
-export async function issueCredentialsByClient(
-    clientId: string, clientSecret: string, state: string, scopes: string[]
-): Promise<ICredentials> {
-    // クライアントの存在確認
-    const clientAdapter = sskts.adapter.client(sskts.mongoose.connection);
-    const clientDoc = await clientAdapter.clientModel.findById(clientId).exec();
-    if (clientDoc === null) {
-        throw new Error(MESSAGE_CLIENT_NOT_FOUND);
-    }
+// export async function issueCredentialsByClient(
+//     clientId: string, clientSecret: string, state: string, scopes: string[]
+// ): Promise<ICredentials> {
+//     // クライアントの存在確認
+//     const clientAdapter = sskts.adapter.client(sskts.mongoose.connection);
+//     const clientDoc = await clientAdapter.clientModel.findById(clientId).exec();
+//     if (clientDoc === null) {
+//         throw new Error(MESSAGE_CLIENT_NOT_FOUND);
+//     }
 
-    // クライアントシークレット`検証
-    debug('validating client secret...', clientSecret);
-    if (!await bcrypt.compare(clientSecret, clientDoc.get('secret_hash'))) {
-        throw new Error(MESSAGE_INVALID_CLIENT_CREDENTIALS);
-    }
+//     // クライアントシークレット`検証
+//     debug('validating client secret...', clientSecret);
+//     if (!await bcrypt.compare(clientSecret, clientDoc.get('secret_hash'))) {
+//         throw new Error(MESSAGE_INVALID_CLIENT_CREDENTIALS);
+//     }
 
-    const payload = sskts.factory.clientUser.create({
-        client: clientId,
-        state: state,
-        scopes: scopes
-    });
+//     const payload = sskts.factory.clientUser.create({
+//         client: clientId,
+//         state: state,
+//         scopes: scopes
+//     });
 
-    return await payload2credentials(payload);
-}
+//     return await payload2credentials(payload);
+// }
 
 /**
  * Googleのid tokenから資格情報を発行する
  */
-export async function issueCredentialsByGoogleToken(
-    idToken: string, clientId: string, state: string, scopes: string[]
-): Promise<ICredentials> {
-    // クライアントの存在確認
-    const clientAdapter = sskts.adapter.client(sskts.mongoose.connection);
-    const clientDoc = await clientAdapter.clientModel.findById(clientId, '_id').exec();
-    if (clientDoc === null) {
-        throw new Error(MESSAGE_CLIENT_NOT_FOUND);
-    }
+// export async function issueCredentialsByGoogleToken(
+//     idToken: string, clientId: string, state: string, scopes: string[]
+// ): Promise<ICredentials> {
+//     // クライアントの存在確認
+//     const clientAdapter = sskts.adapter.client(sskts.mongoose.connection);
+//     const clientDoc = await clientAdapter.clientModel.findById(clientId, '_id').exec();
+//     if (clientDoc === null) {
+//         throw new Error(MESSAGE_CLIENT_NOT_FOUND);
+//     }
 
-    const userInfo = await verifyGoogleIdToken(idToken);
+//     const userInfo = await verifyGoogleIdToken(idToken);
 
-    // 会員検索(なければ登録)
-    const personAdapter = await sskts.adapter.person(sskts.mongoose.connection);
-    const person = {
-        typeOf: 'Person',
-        email: userInfo.email,
-        givenName: '',
-        familyName: '',
-        telephone: '',
-        memberOf: {
-            openId: {
-                provider: userInfo.iss,
-                userId: userInfo.sub
-            },
-            hostingOrganization: {},
-            membershipNumber: `${userInfo.iss}-${userInfo.sub}`,
-            programName: 'シネマサンシャインプレミアム'
-        }
-    };
-    const personDoc = await personAdapter.personModel.findOneAndUpdate(
-        {
-            'memberOf.openId.provider': person.memberOf.openId.provider,
-            'memberOf.openId.userId': person.memberOf.openId.userId
-        },
-        {
-            $setOnInsert: person // 新規の場合のみ更新
-        },
-        { new: true, upsert: true }
-    ).exec();
+//     // 会員検索(なければ登録)
+//     const personAdapter = await sskts.adapter.person(sskts.mongoose.connection);
+//     const person = {
+//         typeOf: 'Person',
+//         email: userInfo.email,
+//         givenName: '',
+//         familyName: '',
+//         telephone: '',
+//         memberOf: {
+//             openId: {
+//                 provider: userInfo.iss,
+//                 userId: userInfo.sub
+//             },
+//             hostingOrganization: {},
+//             membershipNumber: `${userInfo.iss}-${userInfo.sub}`,
+//             programName: 'シネマサンシャインプレミアム'
+//         }
+//     };
+//     const personDoc = await personAdapter.personModel.findOneAndUpdate(
+//         {
+//             'memberOf.openId.provider': person.memberOf.openId.provider,
+//             'memberOf.openId.userId': person.memberOf.openId.userId
+//         },
+//         {
+//             $setOnInsert: person // 新規の場合のみ更新
+//         },
+//         { new: true, upsert: true }
+//     ).exec();
 
-    if (personDoc === null) {
-        throw new Error('member not found');
-    }
+//     if (personDoc === null) {
+//         throw new Error('member not found');
+//     }
 
-    const payload = sskts.factory.clientUser.create(<any>{
-        client: clientId,
-        person: personDoc.toObject(),
-        state: state,
-        scopes: scopes
-    });
+//     const payload = sskts.factory.clientUser.create(<any>{
+//         client: clientId,
+//         person: personDoc.toObject(),
+//         state: state,
+//         scopes: scopes
+//     });
 
-    return await payload2credentials(payload);
-}
+//     return await payload2credentials(payload);
+// }
 
-async function verifyGoogleIdToken(idToken: string) {
-    return new Promise<any>((resolve, reject) => {
+// async function verifyGoogleIdToken(idToken: string) {
+//     return new Promise<any>((resolve, reject) => {
 
-        // idTokenをGoogleで検証
-        const CLIENT_ID = '932934324671-66kasujntj2ja7c5k4k55ij6pakpqir4.apps.googleusercontent.com';
-        const auth = new googleAuth();
-        const client = new auth.OAuth2(CLIENT_ID, '', '');
-        client.verifyIdToken(
-            idToken,
-            CLIENT_ID,
-            // Or, if multiple clients access the backend:
-            //[CLIENT_ID_1, CLIENT_ID_2, CLIENT_ID_3],
-            async (error: any, login: any) => {
-                if (error !== null) {
-                    reject(error);
+//         // idTokenをGoogleで検証
+//         const CLIENT_ID = '932934324671-66kasujntj2ja7c5k4k55ij6pakpqir4.apps.googleusercontent.com';
+//         const auth = new googleAuth();
+//         const client = new auth.OAuth2(CLIENT_ID, '', '');
+//         client.verifyIdToken(
+//             idToken,
+//             CLIENT_ID,
+//             // Or, if multiple clients access the backend:
+//             //[CLIENT_ID_1, CLIENT_ID_2, CLIENT_ID_3],
+//             async (error: any, login: any) => {
+//                 if (error !== null) {
+//                     reject(error);
 
-                    return;
-                }
+//                     return;
+//                 }
 
-                // todo audのチェック
+//                 // todo audのチェック
 
-                const payload = login.getPayload();
-                debug('payload is', payload);
+//                 const payload = login.getPayload();
+//                 debug('payload is', payload);
 
-                // const userId = payload.sub;
-                // If request specified a G Suite domain:
-                //var domain = payload['hd'];
+//                 // const userId = payload.sub;
+//                 // If request specified a G Suite domain:
+//                 //var domain = payload['hd'];
 
-                resolve(payload);
-            });
-    });
-}
+//                 resolve(payload);
+//             });
+//     });
+// }
 
 /**
  * LINEの認可コードから資格情報を発行する
  */
-export async function issueCredentialsByLINEAuthorizationCode(
-    code: string, redirectUri: string, clientId: string, state: string, scopes: string[]
-): Promise<ICredentials> {
-    // クライアントの存在確認
-    const clientAdapter = sskts.adapter.client(sskts.mongoose.connection);
-    const clientDoc = await clientAdapter.clientModel.findById(clientId, '_id').exec();
-    if (clientDoc === null) {
-        throw new Error(MESSAGE_CLIENT_NOT_FOUND);
-    }
+// export async function issueCredentialsByLINEAuthorizationCode(
+//     code: string, redirectUri: string, clientId: string, state: string, scopes: string[]
+// ): Promise<ICredentials> {
+//     // クライアントの存在確認
+//     const clientAdapter = sskts.adapter.client(sskts.mongoose.connection);
+//     const clientDoc = await clientAdapter.clientModel.findById(clientId, '_id').exec();
+//     if (clientDoc === null) {
+//         throw new Error(MESSAGE_CLIENT_NOT_FOUND);
+//     }
 
-    const form = {
-        grant_type: 'authorization_code',
-        client_id: '1527681488',
-        client_secret: 'cf3540f8c004f9e8926a5090ae6a036d',
-        code: code,
-        redirect_uri: redirectUri
-    };
-    debug('getting an access token...', form);
-    const accessToken = await request.post({
-        url: 'https://api.line.me/v2/oauth/accessToken',
-        form: form,
-        json: true,
-        simple: false,
-        resolveWithFullResponse: true,
-        useQuerystring: true
-    }).then((response) => {
-        debug(response.body);
+//     const form = {
+//         grant_type: 'authorization_code',
+//         client_id: '1527681488',
+//         client_secret: 'cf3540f8c004f9e8926a5090ae6a036d',
+//         code: code,
+//         redirect_uri: redirectUri
+//     };
+//     debug('getting an access token...', form);
+//     const accessToken = await request.post({
+//         url: 'https://api.line.me/v2/oauth/accessToken',
+//         form: form,
+//         json: true,
+//         simple: false,
+//         resolveWithFullResponse: true,
+//         useQuerystring: true
+//     }).then((response) => {
+//         debug(response.body);
 
-        return response.body.access_token;
-    });
-    debug('an access token got', accessToken);
+//         return response.body.access_token;
+//     });
+//     debug('an access token got', accessToken);
 
-    // LINEプロフィール取得
-    debug('getting user profiles...');
-    const profile = await request.get({
-        url: 'https://api.line.me/v2/profile',
-        auth: { bearer: accessToken },
-        json: true,
-        simple: false,
-        resolveWithFullResponse: true,
-        useQuerystring: true
-    }).then((response) => {
-        debug(response.body);
+//     // LINEプロフィール取得
+//     debug('getting user profiles...');
+//     const profile = await request.get({
+//         url: 'https://api.line.me/v2/profile',
+//         auth: { bearer: accessToken },
+//         json: true,
+//         simple: false,
+//         resolveWithFullResponse: true,
+//         useQuerystring: true
+//     }).then((response) => {
+//         debug(response.body);
 
-        return response.body;
-    });
-    debug('profile is', profile);
+//         return response.body;
+//     });
+//     debug('profile is', profile);
 
-    // 会員検索(なければ登録)
-    const personAdapter = await sskts.adapter.person(sskts.mongoose.connection);
-    const person = {
-        typeOf: 'Person',
-        email: '',
-        givenName: '',
-        familyName: '',
-        telephone: '',
-        memberOf: {
-            openId: {
-                provider: 'LINE',
-                userId: profile.userId
-            },
-            hostingOrganization: {},
-            membershipNumber: `LINE-${profile.userId}`,
-            programName: 'シネマサンシャインプレミアム'
-        }
-    };
-    const personDoc = await personAdapter.personModel.findOneAndUpdate(
-        {
-            'memberOf.openId.provider': person.memberOf.openId.provider,
-            'memberOf.openId.userId': person.memberOf.openId.userId
-        },
-        {
-            $setOnInsert: person // 新規の場合のみ更新
-        },
-        { new: true, upsert: true }
-    ).exec();
+//     // 会員検索(なければ登録)
+//     const personAdapter = await sskts.adapter.person(sskts.mongoose.connection);
+//     const person = {
+//         typeOf: 'Person',
+//         email: '',
+//         givenName: '',
+//         familyName: '',
+//         telephone: '',
+//         memberOf: {
+//             openId: {
+//                 provider: 'LINE',
+//                 userId: profile.userId
+//             },
+//             hostingOrganization: {},
+//             membershipNumber: `LINE-${profile.userId}`,
+//             programName: 'シネマサンシャインプレミアム'
+//         }
+//     };
+//     const personDoc = await personAdapter.personModel.findOneAndUpdate(
+//         {
+//             'memberOf.openId.provider': person.memberOf.openId.provider,
+//             'memberOf.openId.userId': person.memberOf.openId.userId
+//         },
+//         {
+//             $setOnInsert: person // 新規の場合のみ更新
+//         },
+//         { new: true, upsert: true }
+//     ).exec();
 
-    if (personDoc === null) {
-        throw new Error('member not found');
-    }
+//     if (personDoc === null) {
+//         throw new Error('member not found');
+//     }
 
-    const payload = sskts.factory.clientUser.create(<any>{
-        client: clientId,
-        person: personDoc.toObject(),
-        state: state,
-        scopes: scopes
-    });
+//     const payload = sskts.factory.clientUser.create(<any>{
+//         client: clientId,
+//         person: personDoc.toObject(),
+//         state: state,
+//         scopes: scopes
+//     });
 
-    return await payload2credentials(payload);
-}
+//     return await payload2credentials(payload);
+// }
 
 /**
  * 任意のデータをJWTを使用して資格情報へ変換する